Refactor UnderRowStatusIcon to a shared component
